Add token authorization for delete and patch routes

+++ controllers/blogs.js
@@ -1,6 +1,7 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
 const User = require("../models/user");
+const { userExtractor } = require("../utils/middleware");
 
 blogsRouter.get("/", async (req, res) => {
   const blogs = await Blog.find({}).populate("user", { name: 1, username: 1 });
@@ -19,11 +20,11 @@ blogsRouter.get("/:id", async (req, res) => {
   res.json(blog);
 });
 
-blogsRouter.post("/", async (req, res) => {
+blogsRouter.post("/", userExtractor, async (req, res) => {
   const body = req.body;
-  const user = await User.findById(body.userId);
+  const user = await User.findById(req.user);
   if (!user) {
-    return res.status(400).json({ error: "Invalid user" });
+    return res.status(401).json({ error: "Invalid user" });
   }
   const blog = new Blog({
     title: body.title,
@@ -39,14 +40,25 @@ blogsRouter.post("/", async (req, res) => {
   res.status(201).json(savedBlog);
 });
 
-blogsRouter.delete("/:id", async (req, res) => {
+blogsRouter.delete("/:id", userExtractor, async (req, res) => {
   const id = req.params.id;
+  const blog = await Blog.findById(id);
+  if (blog.user.toString() !== req.user) {
+    return res.status(404).json({ error: "Unauthorized command" });
+  }
   await Blog.findByIdAndDelete(id);
   res.status(204).end();
 });
 
-blogsRouter.patch("/:id", async (req, res) => {
+blogsRouter.patch("/:id", userExtractor, async (req, res) => {
   const id = req.params.id;
+  const blog = await Blog.findById(id);
+  if (!blog) {
+    return res.status(404).json({ error: "Blog not found" });
+  }
+  if (blog.user.toString() !== req.user) {
+    return res.status(404).json({ error: "Unauthorized command" });
+  }
   const newLikes = {
     likes: req.body.likes,
   };
@@ -55,10 +67,6 @@ blogsRouter.patch("/:id", async (req, res) => {
     new: true,
     runValidators: true,
   });
-
-  if (!updatedBlog) {
-    return res.status(404).json({ error: "Blog not found" });
-  }
   res.json(updatedBlog);
 });
